app/requirements.txt
langchain==0.1.20
langchain-openai==0.1.1
langchain-qdrant==0.1.3
fastembed==0.3.4
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.1